Тесты модуля HTTPTRansport
