+++ src/modules/fetch.test.ts
@@ -1,38 +1,23 @@
-import { HTTPTransport } from './fetch'
+import { mockFetch } from '../utils/tests'
+import { HTTPTransport, HTTPTransportResponseError } from './fetch'
 
 describe('HTTP Transport', () => {
   const baseUrl = 'https://test-api.com/v2'
   const fetch = new HTTPTransport(baseUrl)
 
-  function mockFetch<T>(status: number, data?: T) {
-    const xhrMockObj = {
-      open: jest.fn(),
-      send: jest.fn(),
-      setRequestHeader: jest.fn(),
-      readyState: 4,
-      status,
-      response: data,
-    }
-
-    const xhrMockClass = () => xhrMockObj
-
-    // @ts-ignore
-    window.XMLHttpRequest = jest.fn().mockImplementation(xhrMockClass)
-
-    setTimeout(() => {
-      // @ts-ignore
-      xhrMockObj['onreadystatechange']()
-    }, 0)
-
-    return xhrMockClass
-  }
-
-  test('fetch return data', async () => {
+  test('should return response (ok status code)', async () => {
     const data = {
       users: [{ id: 1 }, { id: 2 }],
     }
     const mock = mockFetch(200, data)
     const { response } = await fetch.get('/users')
     expect(data).toEqual(response)
+  })
+
+  test('should return error (error status code)', async () => {
+    const mock = mockFetch(400)
+    await expect(fetch.get('/users')).rejects.toBeInstanceOf(
+      HTTPTransportResponseError
+    )
   })
 })
